display which loot index is next in caches/chests

+++ src/itemstore.c
@@ -119,10 +119,10 @@ int EmptySlotsInStore(t_itemStores store) {
   return retval;
 }
 
-int GetNextIndexInStore(t_itemStores store) {
+size_t GetNextIndexInStore(t_itemStores store) {
   if (stores == NULL) return -1;
 
-  t_itemStore *thisStore = &stores[(int)store];
+  t_itemStore *thisStore = &stores[(size_t)store];
 
   for (int x = 0; x < thisStore->length; x++) {
     if (thisStore->collected[x] == 0) return x;
